Extract reconciler event handler in its own file to get rid of `VisibleForTesting` annotations
